Move methods to test-helpers
